Moved everything back Added more blueprints
