+++ backend/run.py
@@ -1,11 +1,13 @@
-from flask import Flask, current_app, jsonify, render_template, request
-
+from flask import (
+  Flask, current_app, jsonify, render_template, 
+  )
 from flask_cors import CORS
 import os
 import logging
-from . import model
 from . import customJson
-from .controller import dbController
+from .controller import (
+  dbController, viewController,
+  apiController)
 LOG = logging.getLogger(__name__)
 
 #config
@@ -27,57 +29,12 @@ app.json_encoder = customJson.Custom_Jsonify
 CORS(app,resources={r'/*': {'origins':'*'}})
 
 app.register_blueprint(dbController.bp)
-
-"""
-@app.route('/db/init', methods=['GET'])
-def initdb():
-  model.dbModel.init()
-  return jsonify('Initialized database')
-
-@app.route('/db/reset', methods=['GET'])
-def resetdb():
-  model.dbModel.reset()
-  return jsonify('Reset database')
-"""
-@app.route('/db/view/<tableName>', methods=['GET'])
-def viewdb(tableName):
-  res = model.dbModel.view(tableName)
-  app.logger.warn(res)
-  return jsonify(res)
-
-@app.route('/db/view/<tableName>/<number>', methods=['GET'])
-def viewdbrow(tableName, number):
-  res = model.dbModel.viewRow(tableName,number)
-  app.logger.warn(res)
-  return jsonify(res)
-
-"""
-@app.route('/db/add', methods=['GET'])
-def addVal():
-  model.dbModel.insertValues()
-  return jsonify('init table database values')
-
-@app.route('/db/clear', methods=['GET'])
-def cleardb():
-  model.dbModel.removeValues()
-  return jsonify('Clear database values')
-"""
+app.register_blueprint(viewController.bp)
+app.register_blueprint(apiController.bp)
 
 @app.route('/test', methods=['GET'])
 def test():
   return jsonify('help!')
-
-@app.route('/api/getEvents',methods=['POST'])
-def getEvent():
-  clubName = request.get_json().get('nm','')
-  res = model.dbModel.getEventfromClub(clubName)
-  return jsonify(res)
-
-@app.route('/api/getEvents2',methods=['GET'])
-def getEvent2():
-  app.logger.warn(request.args.to_dict())
-  res = model.dbModel.getEventfromClub2(**request.args.to_dict())
-  return jsonify(res)
 
 @app.route('/',defaults={'path': ''},methods=['GET'])
 @app.route('/<path:path>')
